venue-booking-backend/apps/venues/translation.py
from modeltranslation.translator import TranslationOptions, register

from .models import Venue


@register(Venue)
class VenueTranslationOptions(TranslationOptions):
    fields = ("name", "address", "description", "amenities")
